refactor: 拆分 LLM provider 候选链路由实现
将 qq-maid-llm/src/provider/mod.rs 按 LLM 基础设施职责拆分为：
- routing.rs：ModelRouteProvider 与候选链执行、provider 选择
- stream_state.rs：流式候选链 RouteStreamState 与 next_route_stream_event 状态推进
- route_config.rs：auto_*、ensure_*_api_key、route_uses_provider 等 build_provider 预检
- route_error.rs：ModelAttemptFailure、aggregate_route_error、错误分类 helper
mod.rs 继续作为 crate 公开入口，保留 LlmProvider trait、ChatOutcome、ToolChatRequest、
LlmStreamEvent、collect_llm_stream、build_provider 等公开 API 与必要 re-export。

不改变公开类型与候选链 retry/fallback 行为，provider-facing API 对 core 侧调用保持兼容。
