+++ qq-maid-llm/src/provider/mod.rs
@@ -2,14 +2,24 @@
 //!
 //! 定义了统一的 [`LlmProvider`] trait，屏蔽不同 LLM API（OpenAI、DeepSeek、BigModel）的差异。
 //! 同时提供通用模型候选链路由逻辑，以及 [`ChatOutcome`] 等通用类型。
+//!
+//! 本模块作为 provider-facing 的公开入口；模型候选链的执行、流式状态、route 配置预检和
+//! 失败聚合等基础设施分别拆分到 `routing`、`stream_state`、`route_config`、`route_error` 子模块，
+//! 这里仅做组装与 re-export。
 
 pub mod bigmodel;
 pub mod deepseek;
 pub mod limiter;
 pub mod openai;
+mod route_config;
+mod route_error;
+mod routing;
 pub mod status;
+mod stream_state;
 #[cfg(test)]
 pub(crate) mod test_support;
+#[cfg(test)]
+mod tests;
 pub mod types;
 
 use std::{pin::Pin, sync::Arc};
@@ -21,9 +31,26 @@ use crate::{
     config::{LlmConfig, ProviderMode},
     error::LlmError,
     metrics::{LlmMetrics, MetricsRecorder},
-    provider::types::{ChatRequest, ModelId, ModelProvider, ModelRoute, TokenUsage},
+    provider::types::{ChatRequest, ModelProvider, TokenUsage},
     tool::{ToolContext, ToolRegistry},
 };
+
+// 候选链构建与 provider 预检 helper 来源于拆分后的子模块，这里 `use` 进来同时供
+// `build_provider` 与测试模块（`tests` 通过 `use super::*` 引用）复用。
+use route_config::{
+    auto_default_route, auto_provider_routes, ensure_required_api_keys_for_routes,
+    ensure_route_supported, provider_kinds_for_routes,
+};
+use routing::ModelRouteProvider;
+
+// `should_try_next_model` 仅在测试中从 mod 入口处直接断言，运行期由 `routing` /
+// `stream_state` 各自引入，因此用 `cfg(test)` 标注，避免出现 unused import 警告。
+#[cfg(test)]
+use route_error::should_try_next_model;
+// `ModelRoute` 仅在测试中通过 `use super::*` 引用解析模型配置；运行期 `build_provider` 只通过
+// `config.model_route` 字段间接使用该类型，不需要在 mod 入口处直接命名。
+#[cfg(test)]
+use crate::provider::types::ModelRoute;
 
 /// Tool Loop 中单次工具执行的结果摘要。
 ///
@@ -65,7 +92,7 @@ pub struct ToolChatRequest {
     pub tools: ToolRegistry,
     /// 服务端生成的 Tool 执行上下文。
     pub tool_context: ToolContext,
-    /// 最多允许执行的工具调用轮数。
+    /// 最多允许执行工具调用轮数。
     pub max_rounds: usize,
 }
 
@@ -288,643 +315,3 @@ pub fn build_provider(config: &LlmConfig) -> Result<DynLlmProvider, LlmError> {
         }
     }
 }
-
-/// 通用模型候选链提供商。
-///
-/// 先执行 OpenAI/DeepSeek/BigModel 各自内部的 Responses、Chat Completions、空流补非流等
-/// 兼容策略；只有某个候选整体失败且错误允许跨模型降级时，才尝试下一个候选。
-struct ModelRouteProvider {
-    name: &'static str,
-    default_provider: ModelProvider,
-    default_route: ModelRoute,
-    providers: Vec<(ModelProvider, DynLlmProvider)>,
-    model_display: String,
-}
-
-impl ModelRouteProvider {
-    fn new(
-        name: &'static str,
-        default_provider: ModelProvider,
-        default_route: ModelRoute,
-        providers: Vec<(ModelProvider, DynLlmProvider)>,
-    ) -> Result<Self, LlmError> {
-        if providers.is_empty() {
-            return Err(LlmError::config(
-                "no LLM provider is available for model route",
-            ));
-        }
-        let model_display = default_route.display();
-        Ok(Self {
-            name,
-            default_provider,
-            default_route,
-            providers,
-            model_display,
-        })
-    }
-
-    fn provider_for(&self, provider: ModelProvider) -> Option<&DynLlmProvider> {
-        self.providers
-            .iter()
-            .find(|(candidate, _)| *candidate == provider)
-            .map(|(_, provider)| provider)
-    }
-}
-
-#[async_trait]
-impl LlmProvider for ModelRouteProvider {
-    async fn chat(&self, req: ChatRequest) -> Result<ChatOutcome, LlmError> {
-        let route = match req.model.as_deref() {
-            Some(value) => ModelRoute::parse(value, "request")?,
-            None => self.default_route.clone(),
-        };
-        let task = model_task_name(&req);
-        let mut failures = Vec::new();
-
-        for (index, candidate) in route.candidates().iter().enumerate() {
-            let provider_kind = candidate.provider.unwrap_or(self.default_provider);
-            let provider = self.provider_for(provider_kind).ok_or_else(|| {
-                LlmError::config(format!(
-                    "provider `{}` is not available for model candidate `{}`",
-                    provider_kind.as_str(),
-                    candidate.to_request_model()
-                ))
-            })?;
-            let mut candidate_req = req.clone();
-            candidate_req.model = Some(candidate.to_request_model());
-
-            match provider.chat(candidate_req).await {
-                Ok(mut outcome) => {
-                    tracing::info!(
-                        task,
-                        candidate_index = index,
-                        provider = provider_kind.as_str(),
-                        model = %candidate.name,
-                        result = "success",
-                        "model candidate succeeded"
-                    );
-                    // provider 内部兼容 fallback 与跨模型候选降级语义不同；这里只在
-                    // 真正使用后续模型候选时标记，保持原有候选链行为不变。
-                    outcome.fallback_used |= index > 0;
-                    return Ok(outcome);
-                }
-                Err(err) => {
-                    let fallback = index + 1 < route.len() && should_try_next_model(&err);
-                    tracing::warn!(
-                        task,
-                        candidate_index = index,
-                        provider = provider_kind.as_str(),
-                        model = %candidate.name,
-                        result = "failed",
-                        error_code = err.code.as_str(),
-                        error_stage = err.stage.as_str(),
-                        error_kind = model_error_kind(&err),
-                        fallback,
-                        "model candidate failed"
-                    );
-                    if !fallback {
-                        if route.len() == 1 || !should_try_next_model(&err) {
-                            return Err(err);
-                        }
-                        failures.push(ModelAttemptFailure::new(
-                            index,
-                            provider_kind,
-                            candidate,
-                            err,
-                        ));
-                        return Err(aggregate_route_error(task, failures));
-                    }
-                    failures.push(ModelAttemptFailure::new(
-                        index,
-                        provider_kind,
-                        candidate,
-                        err,
-                    ));
-                }
-            }
-        }
-
-        Err(aggregate_route_error(task, failures))
-    }
-
-    async fn chat_with_tools(&self, req: ToolChatRequest) -> Result<ChatOutcome, LlmError> {
-        let candidates = match req.chat.model.as_deref() {
-            Some(value) => ModelRoute::parse(value, "request")?.candidates().to_vec(),
-            None => self.default_route.candidates().to_vec(),
-        };
-        let Some(candidate) = candidates.first() else {
-            return Err(LlmError::new(
-                "bad_request",
-                "model candidate list must not be empty",
-                "request",
-            ));
-        };
-        let provider_kind = candidate.provider.unwrap_or(self.default_provider);
-        let Some(provider) = self.provider_for(provider_kind).cloned() else {
-            return Err(LlmError::config(format!(
-                "no provider configured for {}",
-                provider_kind.as_str()
-            )));
-        };
-        let model = candidate.to_request_model();
-        let mut chat = req.chat;
-        chat.model = Some(model.clone());
-        // Tool Loop 期间固定首个候选和 provider；未适配 Tool Calling 的 provider
-        // 安全回退同一候选的普通聊天，不进入后续候选链。
-        if provider.tool_calling_protocol(Some(&model)).is_none() {
-            return provider.chat(chat).await;
-        }
-        provider
-            .chat_with_tools(ToolChatRequest {
-                chat,
-                tools: req.tools,
-                tool_context: req.tool_context,
-                max_rounds: req.max_rounds,
-            })
-            .await
-    }
-
-    fn tool_calling_protocol(&self, model: Option<&str>) -> Option<ToolCallingProtocol> {
-        let candidates = match model {
-            Some(value) => ModelRoute::parse(value, "request")
-                .ok()?
-                .candidates()
-                .to_vec(),
-            None => self.default_route.candidates().to_vec(),
-        };
-        let candidate = candidates.first()?;
-        let provider_kind = candidate.provider.unwrap_or(self.default_provider);
-        let provider = self.provider_for(provider_kind)?;
-        let request_model = candidate.to_request_model();
-        provider.tool_calling_protocol(Some(&request_model))
-    }
-
-    async fn stream_chat(&self, req: ChatRequest) -> Result<LlmStream, LlmError> {
-        let route = match req.model.as_deref() {
-            Some(value) => ModelRoute::parse(value, "request")?,
-            None => self.default_route.clone(),
-        };
-        let task = model_task_name(&req).to_owned();
-        let candidates = route.candidates().to_vec();
-        let providers = self.providers.clone();
-        let default_provider = self.default_provider;
-
-        Ok(Box::pin(stream::unfold(
-            RouteStreamState {
-                req,
-                task,
-                candidates,
-                providers,
-                default_provider,
-                candidate_index: 0,
-                current_stream: None,
-                current_attempt: None,
-                failures: Vec::new(),
-                emitted_non_empty_delta: false,
-                done: false,
-            },
-            |mut state| async move {
-                let event = next_route_stream_event(&mut state).await;
-                event.map(|event| (event, state))
-            },
-        )))
-    }
-
-    fn name(&self) -> &'static str {
-        self.name
-    }
-
-    fn model(&self) -> &str {
-        &self.model_display
-    }
-
-    fn stream_enabled(&self) -> bool {
-        self.providers
-            .first()
-            .map(|(_, provider)| provider.stream_enabled())
-            .unwrap_or(false)
-    }
-}
-
-struct RouteStreamState {
-    req: ChatRequest,
-    task: String,
-    candidates: Vec<ModelId>,
-    providers: Vec<(ModelProvider, DynLlmProvider)>,
-    default_provider: ModelProvider,
-    candidate_index: usize,
-    current_stream: Option<LlmStream>,
-    current_attempt: Option<(usize, ModelProvider, ModelId)>,
-    failures: Vec<ModelAttemptFailure>,
-    emitted_non_empty_delta: bool,
-    done: bool,
-}
-
-async fn next_route_stream_event(
-    state: &mut RouteStreamState,
-) -> Option<Result<LlmStreamEvent, LlmError>> {
-    loop {
-        if state.done {
-            return None;
-        }
-        if state.current_stream.is_none() {
-            match start_next_route_candidate(state).await {
-                Ok(true) => {}
-                Ok(false) => {
-                    state.done = true;
-                    return Some(Err(aggregate_route_error(
-                        &state.task,
-                        std::mem::take(&mut state.failures),
-                    )));
-                }
-                Err(err) => {
-                    state.done = true;
-                    return Some(Err(err));
-                }
-            }
-        }
-
-        let Some(stream) = state.current_stream.as_mut() else {
-            continue;
-        };
-        match stream.next().await {
-            Some(Ok(LlmStreamEvent::TextDelta(delta))) => {
-                if !delta.is_empty() {
-                    state.emitted_non_empty_delta = true;
-                }
-                return Some(Ok(LlmStreamEvent::TextDelta(delta)));
-            }
-            Some(Ok(LlmStreamEvent::Completed {
-                usage,
-                finish_reason,
-                fallback_used,
-            })) => {
-                if !state.emitted_non_empty_delta {
-                    let err =
-                        LlmError::provider("LLM stream returned empty text output", "provider");
-                    record_current_route_failure(state, err);
-                    state.current_stream = None;
-                    state.current_attempt = None;
-                    continue;
-                }
-                let fallback_used = fallback_used
-                    || state
-                        .current_attempt
-                        .as_ref()
-                        .is_some_and(|(index, _, _)| *index > 0);
-                state.done = true;
-                return Some(Ok(LlmStreamEvent::Completed {
-                    usage,
-                    finish_reason,
-                    fallback_used,
-                }));
-            }
-            Some(Err(err)) => {
-                if state.emitted_non_empty_delta {
-                    state.done = true;
-                    return Some(Err(LlmError::new(
-                        err.code,
-                        err.message,
-                        "stream_after_delta",
-                    )));
-                }
-                record_current_route_failure(state, err);
-                state.current_stream = None;
-                state.current_attempt = None;
-            }
-            None => {
-                if state.emitted_non_empty_delta {
-                    state.done = true;
-                    return Some(Err(LlmError::provider(
-                        "LLM stream ended before completion after emitting text",
-                        "stream_after_delta",
-                    )));
-                }
-                let err = LlmError::provider("LLM stream ended without completion event", "stream");
-                record_current_route_failure(state, err);
-                state.current_stream = None;
-                state.current_attempt = None;
-            }
-        }
-    }
-}
-
-async fn start_next_route_candidate(state: &mut RouteStreamState) -> Result<bool, LlmError> {
-    while state.candidate_index < state.candidates.len() {
-        let index = state.candidate_index;
-        state.candidate_index += 1;
-        let candidate = state.candidates[index].clone();
-        let provider_kind = candidate.provider.unwrap_or(state.default_provider);
-        let provider = state
-            .providers
-            .iter()
-            .find(|(kind, _)| *kind == provider_kind)
-            .map(|(_, provider)| provider.clone())
-            .ok_or_else(|| {
-                LlmError::config(format!(
-                    "provider `{}` is not available for model candidate `{}`",
-                    provider_kind.as_str(),
-                    candidate.to_request_model()
-                ))
-            })?;
-        let mut candidate_req = state.req.clone();
-        candidate_req.model = Some(candidate.to_request_model());
-        match provider.stream_chat(candidate_req).await {
-            Ok(stream) => {
-                tracing::info!(
-                    task = state.task.as_str(),
-                    candidate_index = index,
-                    provider = provider_kind.as_str(),
-                    model = %candidate.name,
-                    result = "stream_started",
-                    "model candidate stream started"
-                );
-                state.current_stream = Some(stream);
-                state.current_attempt = Some((index, provider_kind, candidate));
-                return Ok(true);
-            }
-            Err(err) => {
-                let fallback =
-                    state.candidate_index < state.candidates.len() && should_try_next_model(&err);
-                tracing::warn!(
-                    task = state.task.as_str(),
-                    candidate_index = index,
-                    provider = provider_kind.as_str(),
-                    model = %candidate.name,
-                    result = "failed",
-                    error_code = err.code.as_str(),
-                    error_stage = err.stage.as_str(),
-                    error_kind = model_error_kind(&err),
-                    fallback,
-                    "model candidate stream init failed"
-                );
-                if !fallback {
-                    return Err(err);
-                }
-                state.failures.push(ModelAttemptFailure::new(
-                    index,
-                    provider_kind,
-                    &candidate,
-                    err,
-                ));
-            }
-        }
-    }
-    Ok(false)
-}
-
-fn record_current_route_failure(state: &mut RouteStreamState, err: LlmError) {
-    let Some((index, provider_kind, candidate)) = state.current_attempt.take() else {
-        state.failures.push(ModelAttemptFailure {
-            index: state.candidate_index,
-            provider: state.default_provider,
-            model: "<unknown>".to_owned(),
-            error: err,
-        });
-        return;
-    };
-    let fallback = state.candidate_index < state.candidates.len() && should_try_next_model(&err);
-    tracing::warn!(
-        task = state.task.as_str(),
-        candidate_index = index,
-        provider = provider_kind.as_str(),
-        model = %candidate.name,
-        result = "failed",
-        error_code = err.code.as_str(),
-        error_stage = err.stage.as_str(),
-        error_kind = model_error_kind(&err),
-        fallback,
-        "model candidate stream failed before text delta"
-    );
-    state.failures.push(ModelAttemptFailure::new(
-        index,
-        provider_kind,
-        &candidate,
-        err,
-    ));
-    if !fallback {
-        state.candidate_index = state.candidates.len();
-    }
-}
-
-#[derive(Debug)]
-struct ModelAttemptFailure {
-    index: usize,
-    provider: ModelProvider,
-    model: String,
-    error: LlmError,
-}
-
-impl ModelAttemptFailure {
-    fn new(index: usize, provider: ModelProvider, candidate: &ModelId, error: LlmError) -> Self {
-        Self {
-            index,
-            provider,
-            model: candidate.name.clone(),
-            error,
-        }
-    }
-}
-
-fn auto_default_route(config: &LlmConfig) -> Result<ModelRoute, LlmError> {
-    let mut candidates = config.model_route.candidates().to_vec();
-    // 兼容旧的 `LLM_PROVIDER=auto` 行为：单个 OpenAI/裸主模型在可恢复失败时，
-    // 仍可降级到 `DEEPSEEK_MODEL`。用户显式写多个候选时则严格按配置顺序执行。
-    if candidates.len() == 1
-        && config.deepseek_api_key.is_some()
-        && candidates[0].provider != Some(ModelProvider::DeepSeek)
-    {
-        let deepseek_model = deepseek::deepseek_config_model(&config.deepseek_model)?;
-        candidates.push(ModelId {
-            provider: Some(ModelProvider::DeepSeek),
-            name: deepseek_model,
-        });
-    }
-    ModelRoute::from_candidates(candidates)
-}
-
-fn auto_provider_routes(
-    config: &LlmConfig,
-    default_route: &ModelRoute,
-) -> Result<Vec<(String, ModelRoute)>, LlmError> {
-    let mut routes = config.configured_model_routes.clone();
-    if let Some((_, route)) = routes.iter_mut().find(|(name, _)| *name == "LLM_MODEL") {
-        // provider 初始化必须使用 auto 模式的实际默认链，才能保留单 OpenAI
-        // 主模型自动追加 DeepSeek fallback 的兼容行为。
-        *route = default_route.clone();
-    }
-    Ok(routes)
-}
-
-fn provider_kinds_for_routes(
-    routes: &[(String, ModelRoute)],
-    default_provider: ModelProvider,
-) -> Vec<ModelProvider> {
-    [
-        ModelProvider::OpenAi,
-        ModelProvider::DeepSeek,
-        ModelProvider::BigModel,
-    ]
-    .into_iter()
-    .filter(|provider| {
-        routes
-            .iter()
-            .any(|(_, route)| route_uses_provider(route, *provider, default_provider))
-    })
-    .collect()
-}
-
-fn ensure_deepseek_api_key_for_routes(
-    config: &LlmConfig,
-    routes: &[(String, ModelRoute)],
-) -> Result<(), LlmError> {
-    let uses_deepseek = routes
-        .iter()
-        .filter_map(|(name, route)| {
-            route_uses_provider(route, ModelProvider::DeepSeek, ModelProvider::OpenAi)
-                .then_some(name.as_str())
-        })
-        .collect::<Vec<_>>()
-        .join(", ");
-    if uses_deepseek.is_empty() {
-        return Ok(());
-    }
-
-    let api_key = config.deepseek_api_key.as_ref().ok_or_else(|| {
-        LlmError::config(format!(
-            "DEEPSEEK_API_KEY is required because configured model routes include DeepSeek: {uses_deepseek}"
-        ))
-    })?;
-    if api_key.trim().is_empty() {
-        return Err(LlmError::config(format!(
-            "DEEPSEEK_API_KEY is required because configured model routes include DeepSeek: {uses_deepseek}"
-        )));
-    }
-    Ok(())
-}
-
-fn ensure_bigmodel_api_key_for_routes(
-    config: &LlmConfig,
-    routes: &[(String, ModelRoute)],
-) -> Result<(), LlmError> {
-    let uses_bigmodel = routes
-        .iter()
-        .filter_map(|(name, route)| {
-            route_uses_provider(route, ModelProvider::BigModel, ModelProvider::OpenAi)
-                .then_some(name.as_str())
-        })
-        .collect::<Vec<_>>()
-        .join(", ");
-    if uses_bigmodel.is_empty() {
-        return Ok(());
-    }
-
-    let api_key = config.bigmodel_api_key.as_ref().ok_or_else(|| {
-        LlmError::config(format!(
-            "BIGMODEL_API_KEY is required because configured model routes include BigModel: {uses_bigmodel}"
-        ))
-    })?;
-    if api_key.trim().is_empty() {
-        return Err(LlmError::config(format!(
-            "BIGMODEL_API_KEY is required because configured model routes include BigModel: {uses_bigmodel}"
-        )));
-    }
-    Ok(())
-}
-
-fn ensure_required_api_keys_for_routes(
-    config: &LlmConfig,
-    routes: &[(String, ModelRoute)],
-) -> Result<(), LlmError> {
-    ensure_deepseek_api_key_for_routes(config, routes)?;
-    ensure_bigmodel_api_key_for_routes(config, routes)
-}
-
-fn ensure_route_supported(
-    route: &ModelRoute,
-    supported: ModelProvider,
-    default_provider: ModelProvider,
-    name: &str,
-) -> Result<(), LlmError> {
-    for candidate in route.candidates() {
-        let provider = candidate.provider.unwrap_or(default_provider);
-        if provider != supported {
-            return Err(LlmError::config(format!(
-                "{name} candidate `{}` requires provider `{}`, but LLM_PROVIDER is `{}`",
-                candidate.to_request_model(),
-                provider.as_str(),
-                supported.as_str()
-            )));
-        }
-    }
-    Ok(())
-}
-
-fn route_uses_provider(
-    route: &ModelRoute,
-    provider: ModelProvider,
-    default_provider: ModelProvider,
-) -> bool {
-    route
-        .candidates()
-        .iter()
-        .any(|candidate| candidate.provider.unwrap_or(default_provider) == provider)
-}
-
-/// 判断当前错误是否允许跨候选模型降级。
-///
-/// 这里只接收上游传输、限流、超时、空响应和 provider 协议类失败；配置错误、
-/// 本地请求构造错误和业务参数错误会直接返回，避免把本地问题放大成多次计费请求。
-fn should_try_next_model(err: &LlmError) -> bool {
-    matches!(
-        err.code.as_str(),
-        "timeout" | "provider_error" | "http_error" | "rate_limited" | "upstream_unavailable"
-    )
-}
-
-fn model_error_kind(err: &LlmError) -> &'static str {
-    match err.code.as_str() {
-        "timeout" => "timeout",
-        "http_error" => "http_error",
-        "provider_error" if matches!(err.stage.as_str(), "stream" | "sse") => "stream_error",
-        "provider_error" if err.stage == "json" => "invalid_response",
-        "provider_error" => "provider_error",
-        "rate_limited" => "rate_limited",
-        "upstream_unavailable" => "upstream_unavailable",
-        "bad_request" => "permanent",
-        "config" => "config",
-        _ => "permanent",
-    }
-}
-
-fn model_task_name(req: &ChatRequest) -> &str {
-    req.metadata
-        .get("purpose")
-        .map(String::as_str)
-        .filter(|value| !value.trim().is_empty())
-        .unwrap_or("chat")
-}
-
-fn aggregate_route_error(task: &str, failures: Vec<ModelAttemptFailure>) -> LlmError {
-    let details = failures
-        .into_iter()
-        .map(|failure| {
-            format!(
-                "#{} {}:{} -> {}@{}",
-                failure.index,
-                failure.provider.as_str(),
-                failure.model,
-                failure.error.code,
-                failure.error.stage
-            )
-        })
-        .collect::<Vec<_>>()
-        .join("; ");
-    LlmError::provider(
-        format!("all model candidates failed for task `{task}`: {details}"),
-        "provider_route",
-    )
-}
-
-#[cfg(test)]
-mod tests;
